Fix BIGINT overflow and DELETE endpoint issues
- Change Route.id and Location.routeId from INTEGER to BIGINT to support timestamp values
- Fix DELETE endpoints to accept id from both query params and request body
- Add migration script for existing databases
- Update setup-db.js to use BIGSERIAL and BIGINT

+++ setup-db.js
@@ -43,7 +43,7 @@ async function setupDatabase() {
     console.log('📋 Creating Route table...');
     await sql`
       CREATE TABLE IF NOT EXISTS "Route" (
-        id SERIAL PRIMARY KEY,
+        id BIGSERIAL PRIMARY KEY,
         route TEXT NOT NULL,
         shift TEXT NOT NULL,
         warehouse TEXT NOT NULL,
@@ -59,7 +59,7 @@ async function setupDatabase() {
     await sql`
       CREATE TABLE IF NOT EXISTS "Location" (
         id SERIAL PRIMARY KEY,
-        "routeId" INTEGER NOT NULL REFERENCES "Route"(id) ON DELETE CASCADE,
+        "routeId" BIGINT NOT NULL REFERENCES "Route"(id) ON DELETE CASCADE,
         location TEXT NOT NULL,
         code TEXT,
         no INTEGER,
